Fix MQL5/MQL4 compilation error: cast ATR SL timeframe to correct enum type
The iATR() call for stop loss used InpATRSLTimeframe without a type cast,
causing "cannot convert enum" in MQL5. Added (ENUM_TIMEFRAMES) cast for MQL5
and (int) cast for MQL4, matching the pattern used by indicator ATR calls.

+++ src/lib/mql4-generator/generators/trading.ts
@@ -267,7 +267,9 @@ export function generateStopLossCode(
         )
       );
       // MQL4: iATR returns value directly, no handles/buffers needed
-      code.onTick.push(`double _atrValue = iATR(Symbol(), InpATRSLTimeframe, InpATRPeriod, 0);`);
+      code.onTick.push(
+        `double _atrValue = iATR(Symbol(), (int)InpATRSLTimeframe, InpATRPeriod, 0);`
+      );
       code.onTick.push('if(_atrValue == 0) { Print("ATR value is zero for SL"); return; }');
       code.onTick.push("double slPips = (_atrValue / _Point) * InpATRMultiplier;");
       break;
